requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
Flask-Cors==4.0.0
gunicorn==21.2.0
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
